Renamed root package to 'clokta' per idiomatic python Removed pyyaml dependencies
