+++ clokta/common.py
